Improve Mozo Messages

+++ MozoSDK/Classes/Common/Model/DTO/Message/ConversationMessage.swift
@@ -8,14 +8,14 @@
 import Foundation
 import SwiftyJSON
 public class ConversationMessage: ResponseObjectSerializable {
-    public var id: Int?
+    public var id: Int64?
     public var message: String?
     public var images: [String]?
     public var timeCreatedOn: Int64?
     public var timeRead: Int64?
     public var userSend: Bool? = false
     
-    public init(id: Int?, message: String?, images: [String]?, timeCreatedOn: Int64?, timeRead: Int64?, userSend: Bool?) {
+    public init(id: Int64?, message: String?, images: [String]?, timeCreatedOn: Int64?, timeRead: Int64?, userSend: Bool?) {
         self.id = id
         self.message = message
         self.images = images
@@ -25,7 +25,7 @@ public class ConversationMessage: ResponseObjectSerializable {
     }
     
     public required init?(json: JSON) {
-        self.id = json["id"].int
+        self.id = json["id"].int64
         self.message = json["message"].string
         self.images = json["images"].array?.filter({ $0.string != nil }).map({ $0.string! })
         self.timeCreatedOn = json["timeCreatedOn"].int64
